コンポーネント構成の変更
