+++ src/components/QuestionsResult/QuestionGroupCard.vue
@@ -1,8 +1,10 @@
 <script setup lang="ts">
 import { onMounted, ref } from 'vue'
+import { useDisplay } from 'vuetify'
 import SectionTitle from '@/components/shared/SectionTitle.vue'
 import SectionCard from '@/components/shared/SectionCard.vue'
-import QuestionResult from '@/components/QuestionsResult/QuestionResult.vue'
+import QuestionResultDesktop from '@/components/QuestionsResult/QuestionResultDesktop.vue'
+import QuestionResultMobile from '@/components/QuestionsResult/QuestionResultMobile.vue'
 import { apiClient } from '@/api/apiClient'
 import type { components } from '@/api/schema'
 type QuestionGroup = components['schemas']['QuestionGroupResponse']
@@ -13,6 +15,8 @@ const props = defineProps<{
   participants: string[]
 }>()
 
+const display = useDisplay()
+const isSmAndUp = display.smAndUp
 const answers = ref<Answer[]>([])
 
 const fetchAnswers = async () => {
@@ -30,12 +34,23 @@ onMounted(async () => {
 <template>
   <section-title :title="questionGroup.name" />
   <section-card>
-    <question-result
-      v-for="question in questionGroup.questions"
-      :key="question.id"
-      :question="question"
-      :answers="answers"
-      :participants="participants"
-    />
+    <template v-if="isSmAndUp">
+      <question-result-desktop
+        v-for="question in questionGroup.questions"
+        :key="question.id"
+        :question="question"
+        :answers="answers"
+        :participants="participants"
+      />
+    </template>
+    <template v-else>
+      <question-result-mobile
+        v-for="question in questionGroup.questions"
+        :key="question.id"
+        :question="question"
+        :answers="answers"
+        :participants="participants"
+      />
+    </template>
   </section-card>
 </template>
